requirements.txt
alembic==1.4.2
aniso8601==8.0.0
certifi==2019.11.28
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-Admin==1.5.5
Flask-Cors==3.0.8
Flask-Migrate==2.5.3
Flask-RESTful==0.3.8
Flask-SQLAlchemy==2.4.1
idna==2.9
imutils==0.5.3
itsdangerous==1.1.0
Jinja2==2.11.1
Mako==1.1.2
Markdown==3.2.1
MarkupSafe==1.1.1
monotonic==1.5
numpy==1.18.2
opencv-contrib-python==4.2.0.34
pdfkit==0.6.1
plotly==4.6.0
psutil==5.7.0
python-dateutil==2.8.1
python-dotenv==0.11.0
python-editor==1.0.4
pytz==2019.3
requests==2.23.0
retrying==1.3.3
six==1.14.0
SQLAlchemy==1.3.13
urllib3==1.25.8
Werkzeug==1.0.0
wrapt==1.12.0
WTForms==2.2.1
